- Started implementation of Xdmf writer (http://www.xdmf.org) - Can already write grids. - Support for HDF5 files > 4GB in both ParaView and VisIt. - Distinguishes between light (XML) and heavy (HDF5) data. - No modifications to our HDF5 format necessary but just describe where everything (node coords, connectivity, time steps, field arrays, etc.) can be found using XML files. - In the future:   o Support for reading Xdmf simulation results on-the-fly.   o Support for writing Xdmf results from different processes or threads in parallel.

+++ source/Driver/harmonicDriver.cc
@@ -254,7 +254,7 @@ namespace CoupledField
       Double timePerStep = totalTime / (Double) actFreqStep_;
       Double remainingTime = (numFreq_ - actFreqStep_) * timePerStep;
       pt::ptime now = pt::second_clock::local_time();
-      now += pt::seconds(remainingTime);
+      now += pt::seconds(static_cast<long int>(remainingTime));
       analysis_id_->Get("timePerStep")->SetValue( timePerStep );
       PtrParamNode envNode = info->Get(ParamNode::HEADER)->Get("environment");
       envNode->Get("estimatedEnd")->SetValue(pt::to_simple_string( now ));
