+++ source/Driver/transientdriver.cc
@@ -257,7 +257,7 @@ namespace CoupledField {
       Double timePerStep = totalTime / (Double) count;
       Double remainingTime = (endStep - actTimeStep_) * timePerStep;
       pt::ptime now = pt::second_clock::local_time();
-      now += pt::seconds(remainingTime);
+      now += pt::seconds(static_cast<long int>(remainingTime));
       analysis_id_->Get("timePerStep")->SetValue( timePerStep );
       PtrParamNode envNode = info->Get(ParamNode::HEADER)->Get("environment");
       envNode->Get("estimatedEnd")->SetValue(pt::to_simple_string( now ));
